Move async/await into another file to prevent syntax errors
